Enum id3::ErrorKind

pub enum ErrorKind {
    Io(Error),
    StringDecoding(Vec<u8>),
    NoTag,
    UnsupportedVersion(u8, u8),
    Parsing,
    InvalidInput,
    UnsupportedFeature,
}

Kinds of errors that may occur while performing metadata operations.

Variants
Io(Error)

An error kind indicating that an IO error has occurred. Contains the original io::Error.

StringDecoding(Vec<u8>)

An error kind indicating that a string decoding error has occurred. Contains the invalid bytes.

NoTag

An error kind indicating that the reader does not contain an ID3 tag.

UnsupportedVersion(u8, u8)

An error kind indicating that the reader contains an unsupported ID3 tag version. Contains the major and minor versions that were detected in the tag.

Parsing

An error kind indicating that parsing error has occurred.

InvalidInput

An error kind indicating that some input was invalid.

UnsupportedFeature

An error kind indicating that a feature is not supported.