Enum id3::frame::Content

pub enum Content {
    Text(String),
    ExtendedText(ExtendedText),
    Link(String),
    ExtendedLink(ExtendedLink),
    Comment(Comment),
    Lyrics(Lyrics),
    Picture(Picture),
    Unknown(Vec<u8>),
}

The decoded contents of a frame.

Variants
Text(String)

A value containing the parsed contents of a text frame.

ExtendedText(ExtendedText)

A value containing the parsed contents of a user defined text frame (TXXX).

Link(String)

A value containing the parsed contents of a web link frame.

ExtendedLink(ExtendedLink)

A value containing the parsed contents of a user defined web link frame (WXXX).

Comment(Comment)

A value containing the parsed contents of a comment frame (COMM).

Lyrics(Lyrics)

A value containing the parsed contents of a lyrics frame (USLT).

Picture(Picture)

A value containing the parsed contents of a picture frame (APIC).

Unknown(Vec<u8>)

A value containing the bytes of a unknown frame.

Methods

impl Content

fn text(&self) -> &String

Returns the Text. Panics if the value is not Text.

fn extended_text(&self) -> &ExtendedText

Returns the ExtendedText. Panics if the value is not ExtendedText.

fn link(&self) -> &String

Returns the Link. Panics if the value is not Link.

fn extended_link(&self) -> &ExtendedLink

Returns the ExtendedLink. Panics if the value is not ExtendedLink.

fn comment(&self) -> &Comment

Returns the Comment. Panics if the value is not Comment.

fn lyrics(&self) -> &Lyrics

Returns the Lyrics. Panics if the value is not Lyrics.

fn picture(&self) -> &Picture

Returns the Picture. Panics if the value is not Picture.

fn unknown(&self) -> &[u8]

Returns the Unknown. Panics if the value is not Unknown.