Struct id3::frame::Flags

pub struct Flags {
    pub tag_alter_preservation: bool,
    pub file_alter_preservation: bool,
    pub read_only: bool,
    pub compression: bool,
    pub encryption: bool,
    pub grouping_identity: bool,
    pub unsynchronization: bool,
    pub data_length_indicator: bool,
}

Flags used in ID3 frames.

Fields
tag_alter_preservation: bool

Indicates whether or not this frame should be discarded if the tag is altered. A value of true indicates the frame should be discarded.

file_alter_preservation: bool

Indicates whether or not this frame should be discarded if the file is altered. A value of true indicates the frame should be discarded.

read_only: bool

Indicates whether or not this frame is intended to be read only.

compression: bool

Indicates whether or not the frame is compressed using zlib. If set 4 bytes for "decompressed size" are appended to the header.

encryption: bool

Indicates whether or not the frame is encrypted. If set a byte indicating which encryption method was used will be appended to the header.

grouping_identity: bool

Indicates whether or not the frame belongs in a group with other frames. If set a group identifier byte is appended to the header.

unsynchronization: bool

This flag indicates whether or not unsynchronisation was applied to this frame.

data_length_indicator: bool

This flag indicates that a data length indicator has been added to the frame.

Methods

impl Flags

fn new() -> Flags

Returns a new Flags with all flags set to false.

fn to_bytes(&self, version: u8) -> Vec<u8>

Returns a vector representation suitable for writing to a file containing an ID3 tag of the specified version.