Struct id3::header::Header

pub struct Header {
    // some fields omitted
}

A type representing the header of an ID3v2 tag.

Reference

ID3v2.3 Informal Standard (Section 3.1)

Methods

impl Header

fn from_reader<R: Read>(reader: &mut R) -> Result<Header>

Construct a new header from a reader.

Guarantees

This function will only ever read 10 bytes from the reader.

Errors

If there is an error reading the bytes from the reader, then this function will return Error::Io.

If the file identifier is incorrect (must be "ID3"), then this function will return Error::InvalidIdentifier.

If there is an invalid version (neither versions can be 255), then this function will return Error::InvalidVersion.

If there is an invalid size (it must be greater than 0 and less than 268435456), then this function will return Error::InvalidSize.

If there is an unrecognized flag, then this function will return Error::UnknownFlag.

fn identifier(&self) -> &[u8; 3]

Gets the file identifier (currently always "ID3").

fn version(&self) -> &Version

Gets the version of the ID3v2 tag.

fn size(&self) -> u32

Gets the size of the tag (not including the header or footer). This will always be greater than 0 and less than 268435456.

fn is_unsynchronized(&self) -> bool

Gets whether or not unsynchronization is used.

fn has_extended_header(&self) -> bool

Gets whether or not the header is followed by an extended header.

fn is_experimental(&self) -> bool

Gets whether or not the tag is in an experimental stage.

fn has_footer(&self) -> bool

Gets whether or not the tag has a footer.

Since this library only supports v2.3.0, a v2.4.0 footer will never be parsed, but we need to know one exists so it can be safely skipped.