Struct id3::Error

pub struct Error {
    pub kind: ErrorKind,
    pub description: &'static str,
}

A structure able to represent any error that may occur while performing metadata operations.

Fields

kind

The kind of error.

description

A human readable string describing the error.

Methods

impl Error

fn new(kind: ErrorKind, description: &'static str) -> Error

Creates a new Error using the error kind and description.