Struct id3::Timestamp

pub struct Timestamp {
    pub year: i32,
    pub month: Option<u8>,
    pub day: Option<u8>,
    pub hour: Option<u8>,
    pub minute: Option<u8>,
    pub second: Option<u8>,
}

Represents a date and time according to the ID3v2.4 spec:

The timestamp fields are based on a subset of ISO 8601. When being as precise as possible the format of a time string is yyyy-MM-ddTHH:mm:ss (year, "-", month, "-", day, "T", hour (out of 24), ":", minutes, ":", seconds), but the precision may be reduced by removing as many time indicators as wanted. Hence valid timestamps are yyyy, yyyy-MM, yyyy-MM-dd, yyyy-MM-ddTHH, yyyy-MM-ddTHH:mm and yyyy-MM-ddTHH:mm:ss. All time stamps are UTC.

Fields
year: i32 month: Option<u8> day: Option<u8> hour: Option<u8> minute: Option<u8> second: Option<u8>
Methods

impl Timestamp

pub fn parse(source: &str) -> Option<Timestamp>

Deprecated
: Use "yyyy".parse()

Parses a timestamp.

pub fn to_string(&self) -> String

Encodes the timestamp for storing in a frame.